Fix PS1
I recently switched my `PS1` to double quotes, which caused it to be
expanded when the prompt is defined rather than when it's displayed.
That prevented functionality like `__git_ps1` from working as expected.
This reverses that change and updates the project to consistently use
single quotes where possible.

Related:

https://www.gnu.org/software/bash/manual/html_node/Quoting.html